+++ apps/laboratory/src/components/Wagmi/WagmiRequestPermissionsSyncTest.tsx
@@ -1,37 +1,32 @@
 import { Button, Stack, Text } from '@chakra-ui/react'
-import { useAccount } from 'wagmi'
-import { walletActionsErc7715 } from 'viem/experimental'
-import { useCallback, useState } from 'react'
+import { useCallback, useMemo, useState } from 'react'
 import { useChakraToast } from '../Toast'
-import { createWalletClient, custom, type Address, type Chain } from 'viem'
-import { EIP_7715_RPC_METHODS } from '../../utils/EIP5792Utils'
+import { toHex, type Address } from 'viem'
 import { usePasskey } from '../../context/PasskeyContext'
-import {
-  useWagmiAvailableCapabilities,
-  type Provider
-} from '../../hooks/useWagmiActiveCapabilities'
 import { useERC7715Permissions } from '../../hooks/useERC7715Permissions'
 import { bigIntReplacer } from '../../utils/CommonUtils'
 import { getPurchaseDonutPermissions } from '../../utils/ERC7715Utils'
 import { serializePublicKey, type P256Credential } from 'webauthn-p256'
-import { KeyTypes } from '../../utils/EncodingUtils'
-import { useAppKitAccount } from '@reown/appkit/react'
+import { useAppKitAccount, useAppKitNetwork } from '@reown/appkit/react'
+import {
+  grantPermissions,
+  isSmartSessionSupported,
+  type SmartSessionGrantPermissionsRequest
+} from '@reown/appkit-experimental/smart-session'
 
 export function WagmiRequestPermissionsSyncTest() {
-  const { provider, supported } = useWagmiAvailableCapabilities({
-    method: EIP_7715_RPC_METHODS.WALLET_GRANT_PERMISSIONS
-  })
-  const { address, isConnected } = useAppKitAccount()
-  const { chain } = useAccount()
+  const { address, isConnected, status } = useAppKitAccount()
+  const { chainId } = useAppKitNetwork()
+  const isSupported = useMemo(() => isSmartSessionSupported(), [status])
 
-  if (!isConnected || !provider || !address || !chain) {
+  if (!isConnected || !address || !chainId) {
     return (
       <Text fontSize="md" color="yellow">
         Wallet not connected
       </Text>
     )
   }
-  if (!supported) {
+  if (!isSupported) {
     return (
       <Text fontSize="md" color="yellow">
         Wallet does not support wallet_grantPermissions rpc method
@@ -39,21 +34,19 @@ export function WagmiRequestPermissionsSyncTest() {
     )
   }
 
-  return <ConnectedTestContent chain={chain} provider={provider} address={address as Address} />
+  return <ConnectedTestContent chainId={chainId} address={address as Address} />
 }
 
 function ConnectedTestContent({
-  chain,
-  provider,
+  chainId,
   address
 }: {
-  chain: Chain
-  provider: Provider
+  chainId: string | number
   address: Address | undefined
 }) {
   const [isRequestPermissionLoading, setRequestPermissionLoading] = useState<boolean>(false)
   const { passkey } = usePasskey()
-  const { grantedPermissions, clearGrantedPermissions, grantPermissions } = useERC7715Permissions()
+  const { clearSmartSession, setSmartSession, smartSession } = useERC7715Permissions()
   const toast = useChakraToast()
 
   const onRequestPermissions = useCallback(async () => {
@@ -65,14 +58,6 @@ function ConnectedTestContent({
       if (!passkey) {
         throw new Error('Passkey not available')
       }
-      if (!provider) {
-        throw new Error('No Provider available, Please connect your wallet.')
-      }
-      const walletClient = createWalletClient({
-        account: address,
-        chain,
-        transport: custom(provider)
-      }).extend(walletActionsErc7715())
       let p256Credential = passkey as P256Credential
       p256Credential = {
         ...p256Credential,
@@ -83,19 +68,34 @@ function ConnectedTestContent({
         }
       }
       const passkeyPublicKey = serializePublicKey(p256Credential.publicKey, { to: 'hex' })
-
       const purchaseDonutPermissions = getPurchaseDonutPermissions()
-      const response = await grantPermissions(walletClient, {
-        permissions: purchaseDonutPermissions,
-        signerKey: {
-          key: passkeyPublicKey,
-          type: KeyTypes.secp256r1
-        }
+      const grantPurchaseDonutPermissions: SmartSessionGrantPermissionsRequest = {
+        // Adding 24 hours to the current time
+        expiry: Math.floor(Date.now() / 1000) + 24 * 60 * 60,
+        chainId: toHex(chainId),
+        address,
+        signer: {
+          type: 'keys',
+          data: {
+            keys: [
+              {
+                type: 'secp256r1',
+                publicKey: passkeyPublicKey
+              }
+            ]
+          }
+        },
+        ...purchaseDonutPermissions
+      }
+      const response = await grantPermissions(grantPurchaseDonutPermissions)
+      setSmartSession({
+        type: 'sync',
+        grantedPermissions: response
       })
       toast({
         type: 'success',
         title: 'Permissions Granted',
-        description: JSON.stringify(response.approvedPermissions, bigIntReplacer)
+        description: JSON.stringify(response, bigIntReplacer)
       })
     } catch (error) {
       toast({
@@ -106,22 +106,22 @@ function ConnectedTestContent({
     } finally {
       setRequestPermissionLoading(false)
     }
-  }, [passkey, provider, address, chain, grantPermissions, toast])
+  }, [passkey, address, chainId, grantPermissions, toast])
 
   return (
     <Stack direction={['column', 'column', 'row']}>
       <Button
         data-testid="request-permissions-button"
         onClick={onRequestPermissions}
-        isDisabled={Boolean(isRequestPermissionLoading || Boolean(grantedPermissions))}
+        isDisabled={Boolean(isRequestPermissionLoading || Boolean(smartSession?.type === 'sync'))}
         isLoading={isRequestPermissionLoading}
       >
         Request Permissions
       </Button>
       <Button
-        data-test-id="clear-permissions-button"
-        onClick={clearGrantedPermissions}
-        isDisabled={!grantedPermissions}
+        data-test-id="clear-smart-session-button"
+        onClick={clearSmartSession}
+        isDisabled={!smartSession || smartSession.type !== 'sync'}
       >
         Clear Permissions
       </Button>