+++ apps/laboratory/src/utils/SentryUtil.ts
@@ -1,12 +1,12 @@
 import * as Sentry from '@sentry/react'
-import { BrowserTracing } from '@sentry/browser'
+import { browserTracingIntegration } from '@sentry/browser'
 
 export function bootstrapSentry() {
   const dsn = process.env['NEXT_PUBLIC_SENTRY_DSN']
   if (dsn) {
     Sentry.init({
       dsn,
-      integrations: [new BrowserTracing()],
+      integrations: [browserTracingIntegration()],
       tracesSampleRate: 1.0
     })
   }