+++ apps/laboratory/src/utils/UserOpBuilderServiceUtils.ts
@@ -1,19 +1,29 @@
-import axios, { AxiosError } from 'axios'
 import { bigIntReplacer } from '../utils/CommonUtils'
 import type { Address, Hex } from 'viem'
 import { USEROP_BUILDER_SERVICE_BASE_URL } from './ConstantsUtil'
 
 export type Call = { to: Address; value: bigint; data: Hex }
 
-export type BuildUserOpRequestArguments = {
-  account: Address
-  chainId: number
-  calls: Call[]
-  capabilities: {
-    paymasterService?: { url: string }
-    permissions?: { context: Hex }
-  }
+export type UserOperationWithBigIntAsHex = {
+  sender: Address
+  nonce: Hex
+  factory: Address | undefined
+  factoryData: Hex | undefined
+  callData: Hex
+  callGasLimit: Hex
+  verificationGasLimit: Hex
+  preVerificationGas: Hex
+  maxFeePerGas: Hex
+  maxPriorityFeePerGas: Hex
+  paymaster: Address | undefined
+  paymasterVerificationGasLimit: Hex | undefined
+  paymasterPostOpGasLimit: Hex | undefined
+  paymasterData: Hex | undefined
+  signature: Hex
+  initCode?: never
+  paymasterAndData?: never
 }
+
 /**
  * UserOperation v0.7
  */
@@ -37,25 +47,45 @@ export type UserOperation = {
   paymasterAndData?: never
 }
 
-export type FillUserOpResponse = {
-  userOp: UserOperation
-  hash: Hex
-}
-
 export type ErrorResponse = {
   message: string
   error: string
 }
 
-export type SendUserOpWithSignatureParams = {
-  chainId: number
-  userOp: UserOperation
-  signature: Hex
-  permissionsContext?: Hex
+export type PrepareCallsParams = {
+  from: `0x${string}`
+  chainId: `0x${string}`
+  calls: {
+    to: `0x${string}`
+    data: `0x${string}`
+    value: `0x${string}`
+  }[]
+  capabilities: Record<string, unknown>
 }
-export type SendUserOpWithSignatureResponse = {
-  receipt: Hex
+
+export type PrepareCallsReturnValue = {
+  preparedCalls: {
+    type: string
+    data: unknown
+    chainId: `0x${string}`
+  }
+  signatureRequest: {
+    hash: `0x${string}`
+  }
+  context: string
 }
+
+export type SendPreparedCallsParams = {
+  preparedCalls: {
+    type: string
+    data: unknown
+    chainId: `0x${string}`
+  }
+  signature: `0x${string}`
+  context: string
+}
+
+export type SendPreparedCallsReturnValue = string
 
 // Define a custom error type
 export class UserOpBuilderApiError extends Error {
@@ -68,74 +98,68 @@ export class UserOpBuilderApiError extends Error {
   }
 }
 
-// Function to send requests to the CoSigner API
-async function sendUserOpBuilderRequest<
-  TRequest,
-  TResponse extends object,
-  TQueryParams extends Record<string, string> = Record<string, never>
->(args: {
+async function jsonRpcRequest<TParams, TResult>(
+  method: string,
+  params: TParams,
   url: string
-  headers: Record<string, string>
-  data: TRequest
-  queryParams?: TQueryParams
-  transformRequest?: (data: TRequest) => unknown
-}): Promise<TResponse> {
-  const { url, data, queryParams, headers, transformRequest } = args
-  const transformedData = transformRequest ? transformRequest(data) : data
+): Promise<TResult> {
+  const response = await fetch(url, {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json'
+    },
+    body: JSON.stringify(
+      {
+        jsonrpc: '2.0',
+        id: '1',
+        method,
+        params
+      },
+      bigIntReplacer
+    )
+  })
 
-  try {
-    const response = await axios.post<TResponse | ErrorResponse>(url, transformedData, {
-      params: queryParams,
-      headers
-    })
-
-    if ('error' in response.data) {
-      throw new Error(response.data.error)
-    }
-
-    return response.data
-  } catch (error) {
-    if (axios.isAxiosError(error)) {
-      const axiosError = error as AxiosError
-      if (axiosError.response) {
-        throw new UserOpBuilderApiError(
-          axiosError.response.status,
-          JSON.stringify(axiosError.response.data)
-        )
-      } else {
-        throw new UserOpBuilderApiError(500, 'Network error')
-      }
-    }
-    // Re-throw if it's not an Axios error
-    throw error
+  if (!response.ok) {
+    throw new UserOpBuilderApiError(response.status, await response.text())
   }
+
+  const data = await response.json()
+
+  if ('error' in data) {
+    throw new UserOpBuilderApiError(500, JSON.stringify(data.error))
+  }
+
+  return data.result
 }
 
-export async function buildUserOp(args: BuildUserOpRequestArguments): Promise<FillUserOpResponse> {
-  const response = await sendUserOpBuilderRequest<BuildUserOpRequestArguments, FillUserOpResponse>({
-    url: `${USEROP_BUILDER_SERVICE_BASE_URL}/build`,
-    data: args,
-    headers: {
-      'Content-Type': 'application/json'
-    },
-    transformRequest: data => JSON.stringify(data, bigIntReplacer)
-  })
+export async function prepareCalls(args: PrepareCallsParams): Promise<PrepareCallsReturnValue[]> {
+  const projectId = process.env['NEXT_PUBLIC_PROJECT_ID']
+  if (!projectId) {
+    throw new Error('NEXT_PUBLIC_PROJECT_ID is not set')
+  }
 
-  return response
+  const url = `${USEROP_BUILDER_SERVICE_BASE_URL}?projectId=${projectId}`
+
+  return jsonRpcRequest<PrepareCallsParams[], PrepareCallsReturnValue[]>(
+    'wallet_prepareCalls',
+    [args],
+    url
+  )
 }
 
-export async function sendUserOp(args: SendUserOpWithSignatureParams) {
-  const response = await sendUserOpBuilderRequest<
-    SendUserOpWithSignatureParams,
-    SendUserOpWithSignatureResponse
-  >({
-    url: `${USEROP_BUILDER_SERVICE_BASE_URL}/sendUserOp`,
-    data: args,
-    headers: {
-      'Content-Type': 'application/json'
-    },
-    transformRequest: data => JSON.stringify(data, bigIntReplacer)
-  })
+export async function sendPreparedCalls(
+  args: SendPreparedCallsParams
+): Promise<SendPreparedCallsReturnValue[]> {
+  const projectId = process.env['NEXT_PUBLIC_PROJECT_ID']
+  if (!projectId) {
+    throw new Error('NEXT_PUBLIC_PROJECT_ID is not set')
+  }
 
-  return response
+  const url = `${USEROP_BUILDER_SERVICE_BASE_URL}?projectId=${projectId}`
+
+  return jsonRpcRequest<SendPreparedCallsParams[], SendPreparedCallsReturnValue[]>(
+    'wallet_sendPreparedCalls',
+    [args],
+    url
+  )
 }