+++ apps/laboratory/tests/shared/validators/ModalValidator.ts
@@ -84,7 +84,9 @@ export class ModalValidator {
 
   async expectCaipAddressHaveCorrectNetworkId(caipNetworkId: CaipNetworkId) {
     const address = this.page.getByTestId('appkit-caip-address')
-    await expect(address, 'Correct CAIP address should be present').toContainText(caipNetworkId)
+    await expect(address, 'Correct CAIP address should be present').toContainText(
+      caipNetworkId.toString()
+    )
   }
 
   async expectNetwork(network: string) {
@@ -140,6 +142,21 @@ export class ModalValidator {
   async expectNoSocials() {
     const socialList = this.page.getByTestId('wui-list-social')
     await expect(socialList).toBeHidden()
+  }
+
+  async expectAllWallets() {
+    const allWallets = this.page.getByTestId('all-wallets')
+    await expect(allWallets).toBeVisible()
+  }
+
+  async expectNoTryAgainButton() {
+    const secondaryButton = this.page.getByTestId('w3m-connecting-widget-secondary-button')
+    await expect(secondaryButton).toBeHidden()
+  }
+
+  async expectTryAgainButton() {
+    const secondaryButton = this.page.getByTestId('w3m-connecting-widget-secondary-button')
+    await expect(secondaryButton).toBeVisible()
   }
 
   async expectAlertBarText(text: string) {
@@ -235,6 +252,12 @@ export class ModalValidator {
     }
 
     throw new Error('Call status not confirmed')
+  }
+
+  async expectNetworkVisible(name: string) {
+    const network = this.page.getByTestId(`w3m-network-switch-${name}`)
+    await expect(network).toBeVisible()
+    await expect(network).toBeDisabled()
   }
 
   async expectNetworksDisabled(name: string) {