+++ packages/adapters/ethers/src/tests/client.test.ts
@@ -5,15 +5,28 @@ import { mockOptions } from './mocks/Options'
 import { mockCreateEthersConfig } from './mocks/EthersConfig'
 import mockAppKit from './mocks/AppKit'
 import { mockAuthConnector } from './mocks/AuthConnector'
-import { EthersHelpersUtil, type ProviderId, type ProviderType } from '@reown/appkit-utils/ethers'
-import { ConstantsUtil } from '@reown/appkit-utils'
-import { arbitrum, mainnet, polygon } from '@reown/appkit/networks'
-import { ProviderUtil } from '@reown/appkit/store'
+import { EthersHelpersUtil, type ProviderType } from '@reown/appkit-utils/ethers'
+import { CaipNetworksUtil, ConstantsUtil } from '@reown/appkit-utils'
+import {
+  arbitrum as AppkitArbitrum,
+  mainnet as AppkitMainnet,
+  polygon as AppkitPolygon,
+  optimism as AppkitOptimism,
+  bsc as AppkitBsc
+} from '@reown/appkit/networks'
+import { ProviderUtil, type ProviderIdType } from '@reown/appkit/store'
 import { SafeLocalStorage, SafeLocalStorageKeys } from '@reown/appkit-common'
 import { type BlockchainApiLookupEnsName } from '@reown/appkit'
 import { InfuraProvider, JsonRpcProvider } from 'ethers'
 
 import type { CaipNetwork, ChainNamespace } from '@reown/appkit-common'
+
+const [mainnet, arbitrum, polygon, optimism, bsc] = CaipNetworksUtil.extendCaipNetworks(
+  [AppkitMainnet, AppkitArbitrum, AppkitPolygon, AppkitOptimism, AppkitBsc],
+  { customNetworkImageUrls: {}, projectId: '1234' }
+) as [CaipNetwork, CaipNetwork, CaipNetwork, CaipNetwork, CaipNetwork]
+
+const caipNetworks = [mainnet, arbitrum, polygon] as [CaipNetwork, ...CaipNetwork[]]
 
 vi.mock('@reown/appkit-wallet', () => ({
   W3mFrameProvider: vi.fn().mockImplementation(() => mockAuthConnector),
@@ -26,13 +39,16 @@ vi.mock('@reown/appkit-wallet', () => ({
   }
 }))
 
-vi.mock('@reown/appkit-utils', () => {
+vi.mock('@reown/appkit-utils', async importOriginal => {
+  const actual = await importOriginal()
   const INJECTED_CONNECTOR_ID = 'injected'
   const COINBASE_SDK_CONNECTOR_ID = 'coinbaseWallet'
   const EIP6963_CONNECTOR_ID = 'eip6963'
   const WALLET_CONNECT_CONNECTOR_ID = 'walletConnect'
   const AUTH_CONNECTOR_ID = 'w3mAuth'
   return {
+    // @ts-expect-error - actual is not typed
+    ...actual,
     PresetsUtil: {
       ConnectorTypesMap: {
         [INJECTED_CONNECTOR_ID]: 'INJECTED',
@@ -99,6 +115,8 @@ describe('EthersAdapter', () => {
     client = new EthersAdapter()
     const optionsWithEthersConfig = {
       ...mockOptions,
+      networks: caipNetworks,
+      defaultNetwork: undefined,
       ethersConfig: mockCreateEthersConfig()
     }
     client.construct(mockAppKit, optionsWithEthersConfig)
@@ -115,11 +133,11 @@ describe('EthersAdapter', () => {
     })
 
     it('should set caipNetworks to provided caipNetworks options', () => {
-      expect(client.caipNetworks).toEqual(mockOptions.networks)
+      expect(client.caipNetworks).toEqual(caipNetworks)
     })
 
     it('should set defaultNetwork to first caipNetwork option', () => {
-      expect(client.defaultCaipNetwork).toEqual(mockOptions.networks[0])
+      expect(client.defaultCaipNetwork).toEqual(mainnet)
     })
 
     it('should create ethers config', () => {
@@ -144,16 +162,9 @@ describe('EthersAdapter', () => {
     })
 
     it('should switch network for injected provider', async () => {
-      const newNetwork = {
-        id: 'eip155:137',
-        name: 'Polygon',
-        chainId: '137',
-        rpcUrl: 'https://polygon-rpc.com'
-      } as unknown as CaipNetwork
-
       mockProvider.request.mockResolvedValueOnce(null)
 
-      await client.switchNetwork(newNetwork)
+      await client.switchNetwork(polygon)
 
       expect(mockProvider.request).toHaveBeenCalledWith(
         expect.objectContaining({
@@ -168,18 +179,11 @@ describe('EthersAdapter', () => {
     })
 
     it('should add network if not recognized by wallet', async () => {
-      const newNetwork = {
-        id: 'eip155:42161',
-        name: 'Arbitrum One',
-        chainId: '42161',
-        rpcUrl: 'https://arb1.arbitrum.io/rpc'
-      } as unknown as CaipNetwork
-
       const switchError = { code: 4902 }
       mockProvider.request.mockRejectedValueOnce(switchError)
       mockProvider.request.mockResolvedValueOnce(null)
 
-      await client.switchNetwork(newNetwork)
+      await client.switchNetwork(arbitrum)
 
       expect(mockProvider.request).toHaveBeenCalledWith(
         expect.objectContaining({
@@ -194,17 +198,10 @@ describe('EthersAdapter', () => {
     })
 
     it('should throw error if switching fails', async () => {
-      const newNetwork = {
-        id: 'eip155:56',
-        name: 'Binance Smart Chain',
-        chainId: '56',
-        rpcUrl: 'https://bsc-dataseed.binance.org'
-      } as unknown as CaipNetwork
-
       const switchError = new Error('User rejected the request')
       mockProvider.request.mockRejectedValueOnce(switchError)
 
-      await expect(client.switchNetwork(newNetwork)).rejects.toThrow('Chain is not supported')
+      await expect(client.switchNetwork(bsc)).rejects.toThrow('Chain is not supported')
     })
 
     it('should use universal adapter for WalletConnect', async () => {
@@ -214,18 +211,11 @@ describe('EthersAdapter', () => {
         polkadot: undefined
       })
 
-      const newNetwork = {
-        id: 'eip155:10',
-        name: 'Optimism',
-        chainId: '10',
-        rpcUrl: 'https://mainnet.optimism.io'
-      } as unknown as CaipNetwork
-
-      await client.switchNetwork(newNetwork)
+      await client.switchNetwork(optimism)
 
       expect(
         mockAppKit.universalAdapter?.networkControllerClient.switchCaipNetwork
-      ).toHaveBeenCalledWith(newNetwork)
+      ).toHaveBeenCalledWith(optimism)
     })
 
     it('should set requested CAIP networks for each unique chain namespace', () => {
@@ -758,7 +748,7 @@ describe('EthersAdapter', () => {
 
       await client['syncBalance'](mockAddress, mainnet)
 
-      expect(JsonRpcProvider).toHaveBeenCalledWith(mainnet.rpcUrl, {
+      expect(JsonRpcProvider).toHaveBeenCalledWith(mainnet.rpcUrls.default.http[0], {
         chainId: 1,
         name: 'Ethereum'
       })
@@ -823,7 +813,7 @@ describe('EthersAdapter', () => {
       vi.spyOn(ProviderUtil.state, 'providerIds', 'get').mockReturnValue({
         eip155: ConstantsUtil.EIP6963_CONNECTOR_ID,
         solana: undefined
-      } as Record<ChainNamespace, ProviderId | undefined>)
+      } as Record<ChainNamespace, ProviderIdType | undefined>)
       client['EIP6963Providers'] = [
         {
           info: { name: 'MetaMask', icon: 'icon-url', uuid: 'test-uuid', rdns: 'com.metamask' },
@@ -843,7 +833,7 @@ describe('EthersAdapter', () => {
       vi.spyOn(ProviderUtil.state, 'providerIds', 'get').mockReturnValue({
         eip155: ConstantsUtil.WALLET_CONNECT_CONNECTOR_ID,
         solana: undefined
-      } as Record<ChainNamespace, ProviderId | undefined>)
+      } as Record<ChainNamespace, ProviderIdType | undefined>)
       const mockProvider = {
         session: {
           peer: {
@@ -872,7 +862,7 @@ describe('EthersAdapter', () => {
       vi.spyOn(ProviderUtil.state, 'providerIds', 'get').mockReturnValue({
         eip155: ConstantsUtil.COINBASE_SDK_CONNECTOR_ID,
         solana: undefined
-      } as Record<ChainNamespace, ProviderId | undefined>)
+      } as Record<ChainNamespace, ProviderIdType | undefined>)
       vi.spyOn(mockAppKit, 'getConnectors').mockReturnValue([
         {
           id: ConstantsUtil.COINBASE_SDK_CONNECTOR_ID,