+++ packages/appkit/exports/react.ts
@@ -1,8 +1,8 @@
 import { AppKit } from '../src/client.js'
 import type { AppKitOptions } from '../src/utils/TypesUtil.js'
 import { getAppKit } from '../src/library/react/index.js'
-import packageJson from '../package.json' assert { type: 'json' }
 import { CoreHelperUtil } from '@reown/appkit-core'
+import { PACKAGE_VERSION } from './constants.js'
 
 // -- Views ------------------------------------------------------------
 export * from '@reown/appkit-scaffold-ui'
@@ -27,7 +27,7 @@ export function createAppKit(options: CreateAppKit) {
       sdkVersion: CoreHelperUtil.generateSdkVersion(
         options.adapters ?? [],
         'react',
-        packageJson.version
+        PACKAGE_VERSION
       )
     })
     getAppKit(modal)